fix: Handle attribute references in object declaration subtypes
Previously, when determining the subtype of an object declaration,
only selected components were considered. Now, if the subtype mark
is an attribute reference, the prefix of the attribute is used
instead. This allows the Corresponding_Name_Declaration function
to correctly resolve the subtype even for more complex subtype
expressions involving attributes.